update score in anime entity

+++ src/modules/anime/entities/additional/animeUserData.entity.ts
@@ -1,4 +1,4 @@
-import { Field, Float, Int, ObjectType } from "@nestjs/graphql"
+import { Field, Int, ObjectType } from "@nestjs/graphql"
 import { EListStatus } from "../../../../auxiliary"
 
 @ObjectType()
@@ -6,7 +6,7 @@ class AnimeUserDataModel {
   @Field(() => EListStatus, { nullable: true, defaultValue: null })
   animeStatus: EListStatus
 
-  @Field(() => Float, { nullable: true, defaultValue: null })
+  @Field(() => Int, { nullable: true, defaultValue: null })
   score: number
 
   @Field(() => Int, { nullable: true, defaultValue: null })
